move sensor
