+++ cpp/open3d/Open3D.h.in
@@ -97,12 +97,12 @@
 #include "open3d/visualization/gui/Window.h"
 
 // clang-format off
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/AzureKinectRecorder.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/AzureKinectSensorConfig.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/AzureKinectSensor.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/MKVMetadata.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/MKVReader.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/AzureKinect/MKVWriter.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/RGBDRecorder.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/RGBDSensorConfig.h"
-@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/Sensor/RGBDSensor.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/AzureKinectRecorder.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/AzureKinectSensorConfig.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/AzureKinectSensor.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/MKVMetadata.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/MKVReader.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/AzureKinect/MKVWriter.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/RGBDRecorder.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/RGBDSensorConfig.h"
+@BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/RGBDSensor.h"
