+++ cpp/open3d/io/CMakeLists.txt
@@ -4,7 +4,7 @@ file(GLOB_RECURSE FILE_FORMAT_SOURCE_FILES "FileFormat/*.cpp")
 set(IO_ALL_SOURCE_FILES ${CLASS_IO_SOURCE_FILES} ${FILE_FORMAT_SOURCE_FILES})
 
 if (BUILD_AZURE_KINECT)
-    file(GLOB_RECURSE SENSOR_SOURCE_FILES "Sensor/*.cpp")
+    file(GLOB_RECURSE SENSOR_SOURCE_FILES "sensor/*.cpp")
     set(IO_ALL_SOURCE_FILES ${IO_ALL_SOURCE_FILES} ${SENSOR_SOURCE_FILES})
 endif ()
