+++ cpp/open3d/io/sensor/AzureKinect/PluginMacros.h
