+++ cpp/pybind/io/class_io.cpp
@@ -43,8 +43,8 @@
 #include "pybind/io/io.h"
 
 #ifdef BUILD_AZURE_KINECT
-#include "open3d/io/Sensor/AzureKinect/AzureKinectSensorConfig.h"
-#include "open3d/io/Sensor/AzureKinect/MKVMetadata.h"
+#include "open3d/io/sensor/AzureKinect/AzureKinectSensorConfig.h"
+#include "open3d/io/sensor/AzureKinect/MKVMetadata.h"
 #endif
 
 namespace open3d {
