+++ cpp/pybind/io/sensor.cpp
@@ -25,10 +25,10 @@
 // ----------------------------------------------------------------------------
 
 #include "open3d/geometry/RGBDImage.h"
-#include "open3d/io/Sensor/AzureKinect/AzureKinectRecorder.h"
-#include "open3d/io/Sensor/AzureKinect/AzureKinectSensor.h"
-#include "open3d/io/Sensor/AzureKinect/AzureKinectSensorConfig.h"
-#include "open3d/io/Sensor/AzureKinect/MKVReader.h"
+#include "open3d/io/sensor/AzureKinect/AzureKinectRecorder.h"
+#include "open3d/io/sensor/AzureKinect/AzureKinectSensor.h"
+#include "open3d/io/sensor/AzureKinect/AzureKinectSensorConfig.h"
+#include "open3d/io/sensor/AzureKinect/MKVReader.h"
 
 #include "pybind/docstring.h"
 #include "pybind/io/io.h"
